ckit: correct stale messages about local peer
Although unlikely, it's technically possible for a node to receive a
newer (in terms of the lamport time) gossiped message about itself from
peers.

This can hypothetically happen when the local node restarts and joins a
node in the cluster with an older lamport time.

When this happens, the node can override its own state with the older
message rececived from the peer, bypassing the state change checks.

This commit adds logic to detect whne a node receives a newer message
about itself, and treats it as "stale"; the node will replace the
message with a message of its own for re-broadcasting through the
cluster.